Update real-strawberry-cupcake.css
Changing column-gap to 100px didn't change anything. I give up - removed column-gap altogether to put real-strawberry-cupcake.css back to the way it was. I don't understand why the column rule is screwed up now that the site is being hosted here on GitHub.io. It didn't look that way when it was hosted on ASU's web-hosting. I'll try fixing it more later.